implemented new parameter (-z / --zoom) which controls the render size of the browser. Valid values are 1280 (for 1280/720) 1920 (for 1920/1080) 2560 (for 2560/1440) 3840 (for 3840/2160)

+++ sharedmemory.h
@@ -12,5 +12,5 @@ public:
     ~SharedMemory();
 
     uint8_t* Get();
-    bool Write(uint8_t* data, int size);
+    bool Write(uint8_t* data, size_t size);
 };
